Save stack install script to a file
When run through the CircleCI normal test framework the stack install 
script seems to be missing the last word of the last line (path).

This results in the error:
sh: 804: check_home_local_bin_on_: not found
Exited with code 127

But when runnning interactively or with SSH enabled the problem seems to 
disappear.